+++ plone/app/discussion/browser/javascripts/comments.js
@@ -165,6 +165,15 @@
             return false;
         });
 
+        /**********************************************************************
+         * Edit a comment
+         **********************************************************************/
+	$("form[name='edit']").prepOverlay({
+                cssclass: 'overlay-edit-comment',
+                width: '60%',
+		subtype: 'ajax',
+		filter: '#content>*'
+		})
 
         /**********************************************************************
          * Delete a comment and its answers.